test: use atomic write with hotpatch/compiler test
Some compilers, rely too blindly on chokidar's bad implementation, and as
a result their watch mode is quite broken on Linux and Windows. It will
start compiling with an empty input file when the file is truncated
instead of waiting for actual data being written by fs.writeFileSync.

This fixes that by using atomic writes instead.