lib: add check for script host
All scripts urls not part of the source map were resolved relative to
the cwd so for example:

http://externalhost.com/js/foo.js would resolve to cwd/js/foo.js

This would cause an ENOENT error to be thrown as the directory for that
file could not be found.

This fixes this issue by making sure that only urls with same host as
the page amok is running on are resolved relative to the cwd.